backend/requirements.txt
flask_cors
gevent-websocket
Flask-Sockets
flask==1.1.4
waitress~=2.0.0
argparse~=1.4.0
Werkzeug~=1.0.1
gevent~=21.8.0
pygame~=2.0.1
pyserial~=3.5
svg.path~=4.1
MarkupSafe<2.1.0
git+https://github.com/meerk40t/svgelements
git+https://github.com/mathandy/svgpathtools.git